server/requirements.txt
# FastAPI and core dependencies (updated for Pydantic 2.0 compatibility)
fastapi>=0.100.0
openai>=1.0.0
python-dotenv==1.0.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pypng==0.20220715.0
Pillow>=9.5.0  # Needed for rendering corrected annotation overlays

# Phase 4D.1: Chart Reconstruction
yfinance>=0.2.28
mplfinance>=0.12.10b0
matplotlib>=3.7.0
pandas>=2.0.0

# Phase 5E: Model-Mediated Command Layer
pydantic>=2.0.0

# Phase 5D: Command Router Validation (optional, has fallback)
jsonschema>=4.0.0

# Phase 4A: Database
sqlalchemy>=2.0.0

# Phase 4D: AI Learning System (RAG)
chromadb>=0.4.0